Se ha añadido la funcionalidad de crear conexión con la base de datos e iniciarla. Es posible que en el futuro se requieran otros métodos o modificiaciones.

+++ src/main/kotlin/aemet/mappers/RegistrosMapper.kt
@@ -14,7 +14,7 @@ fun RegistroDto.toRegistro(): Registro{
         horaTempMax = LocalTime.parse(this.horaTempMax) ,
         tempMin = this.tempMin.toDouble(),
         horaTempMin = LocalTime.parse(this.horaTempMin),
-        precipitacion = this.precipitacion.toInt()
+        precipitacion = this.precipitacion.toDouble()
     )
 }
